Edited project settings

+++ Assets/Scripts/Location/Coords/GEO.cs
@@ -2,6 +2,6 @@ using UnityEngine;
 
 public class GEO : MonoBehaviour
 {
-    public float Latitude;
-    public float Longitude;
+    public double Latitude;
+    public double Longitude;
 }
